+++ test/scanning.jl
@@ -4,6 +4,12 @@ using Dates
 using Healpix
 const Sl = Stripeline
 
+# These are the old values that were used by @fincardona to test
+# against the reference position for the Crab Nebula
+const TEST_TENERIFE_LATITUDE_DEG = 28.3
+const TEST_TENERIFE_LONGITUDE_DEG = -16.509722
+const TEST_TENERIFE_HEIGHT_M = 2390
+
 eps = 3e-4 # Corresponds to pointing precision of 1.08 arcseconds
 
 t_start = DateTime(2018, 01, 01, 0, 0, 0)
@@ -18,19 +24,23 @@ crab_dec_astropy_rad = 0.3842255081802917
 crab_position = sqrt(crab_ra_astropy_rad^2 + crab_dec_astropy_rad^2)
 
 # Invert Crab coordinates into telescope pointing directions
-groundq = telescopetoground(_->(0, deg2rad(20), 0), 0)
+groundq = telescopetoground(_ -> (0, deg2rad(20), 0), 0)
 rotmatr = rotationmatrix_normalized(groundq)
 vector = Healpix.ang2vec(dirs...)
 dir = inv(rotmatr) * vector
 
 # Compute skydirs
-(skydirs, skyψ) = genpointings(_->(0, deg2rad(20), 0),
+(skydirs, skyψ) = genpointings(_ -> (0, deg2rad(20), 0),
                                dir, 
                                [0],
                                t_start, 
-                               latitude_deg = TENERIFE_LATITUDE_DEG,
-                               longitude_deg = TENERIFE_LONGITUDE_DEG,
-                               height_m = TENERIFE_HEIGHT_M)
+                               latitude_deg = TEST_TENERIFE_LATITUDE_DEG,
+                               longitude_deg = TEST_TENERIFE_LONGITUDE_DEG,
+                               height_m = TEST_TENERIFE_HEIGHT_M,
+                               precession = true,
+                               nutation = true,
+                               aberration = true,
+                               refraction = true)
 crab_position_skydirs = sqrt(skydirs[1]^2 + skydirs[2]^2)
 @test skydirs[1] ≈ crab_dec_astropy_rad atol = eps
 @test skydirs[2] ≈ crab_ra_astropy_rad atol = eps
@@ -56,13 +66,17 @@ for (idx, day) in enumerate(days)
     vector = Healpix.ang2vec(dirs[idx, 1], dirs[idx, 2])
     dir = inv(rotmatr) * vector
 
-    (skydirections, skyψ) = genpointings(_->(0, deg2rad(20), 0),
+    (skydirections, skyψ) = genpointings(_ -> (0, deg2rad(20), 0),
                                          dir, 
                                          [0], 
                                          day, 
-                                         latitude_deg = TENERIFE_LATITUDE_DEG,
-                                         longitude_deg = TENERIFE_LONGITUDE_DEG,
-                                         height_m = TENERIFE_HEIGHT_M)
+                                         latitude_deg = TEST_TENERIFE_LATITUDE_DEG,
+                                         longitude_deg = TEST_TENERIFE_LONGITUDE_DEG,
+                                         height_m = TEST_TENERIFE_HEIGHT_M,
+                                         precession = true,
+                                         nutation = true,
+                                         aberration = true,
+                                         refraction = true)
 
     skydirs[idx, 1] = skydirections[1]
     skydirs[idx, 2] = skydirections[2]
@@ -92,10 +106,8 @@ times = 0:τ_s:time_duration
 
 (dirs, ψ) = genpointings(db.focalplane["I0"].orientation, 
                          times; 
-                         latitude_deg = TENERIFE_LATITUDE_DEG) do time_s
-    (0,
-                              deg2rad(20.0),
-                              Sl.timetorotang(time_s, spin_velocity))
+                         latitude_deg = TEST_TENERIFE_LATITUDE_DEG) do time_s
+    (0, deg2rad(20.0), Sl.timetorotang(time_s, spin_velocity))
 end
 
 expected_nsamples = convert(Int, time_duration * sampling_rate + 1)
@@ -105,10 +117,8 @@ expected_nsamples = convert(Int, time_duration * sampling_rate + 1)
 (dirs, ψ) = genpointings(db.focalplane["I0"].orientation, 
                          times;
                          ground = true,
-                         latitude_deg = TENERIFE_LATITUDE_DEG) do time_s
-    (0,
-                              deg2rad(20.0),
-                              Sl.timetorotang(time_s, spin_velocity))
+                         latitude_deg = TEST_TENERIFE_LATITUDE_DEG) do time_s
+    (0, deg2rad(20.0), Sl.timetorotang(time_s, spin_velocity))
 end
 
 expected_nsamples = convert(Int, time_duration * sampling_rate + 1)
@@ -118,7 +128,7 @@ expected_nsamples = convert(Int, time_duration * sampling_rate + 1)
 # Check that all the values in the 3rd column are the same
 @test dirs[1:end - 1, 3] ≈ dirs[2:end, 3]
 
-# We strieve for an accuracy of a few arcseconds
+# We strive for an accuracy of a few arcseconds
 @test rad2deg(dirs[1, 3]) ≈ 20.0 atol = 1e-3
 
 ################################################################################
@@ -133,8 +143,18 @@ let defaultdb = InstrumentDB()
     G0_vec = defaultdb.focalplane["G0"].orientation
     V0_vec = defaultdb.focalplane["V0"].orientation
     
-    dirG0, psiG0 = Stripeline.genpointings(wheelfn, G0_vec, timerange)
-    dirV0, psiV0 = Stripeline.genpointings(wheelfn, V0_vec, timerange)
+    dirG0, psiG0 = Stripeline.genpointings(
+        wheelfn,
+        G0_vec,
+        timerange,
+        latitude_deg = TEST_TENERIFE_LATITUDE_DEG,
+    )
+    dirV0, psiV0 = Stripeline.genpointings(
+        wheelfn,
+        V0_vec,
+        timerange,
+        latitude_deg = TEST_TENERIFE_LATITUDE_DEG,
+    )
 
     # We expect G0 to draw larger circles in the sky
     @test minimum(dirG0[:, 1]) < minimum(dirV0[:, 1])